feat(dynamic-table): extract add-column button into Command-based popover
Pull the header "+" column-add UI out of column-manager and records-view
into a dedicated add-column-button with a searchable command popover
(add-column-stack), so both the table toolbar and records header drive
the same flow.
