+++ apps/web/src/components/dynamic-table/dynamic-view.tsx
@@ -283,7 +283,6 @@ export function DynamicView<TData extends object = object>(props: DynamicTablePr
     isLoading = false,
     children,
     emptyState = null,
-    headerActions,
     cellSelection,
     entityLabel,
     onAddNew,
@@ -402,7 +401,6 @@ export function DynamicView<TData extends object = object>(props: DynamicTablePr
       customFilter,
       renderSearch,
       emptyState,
-      headerActions,
       dragDropConfig: props.dragDrop,
       debug: props.debug,
       footerElement,
@@ -432,7 +430,6 @@ export function DynamicView<TData extends object = object>(props: DynamicTablePr
       customFilter,
       renderSearch,
       emptyState,
-      headerActions,
       props.dragDrop,
       props.debug,
       footerElement,
